split reseeds + complete pods test
